Stuck on pvPortMalloc issue
The pvPortMalloc for the image buffer increases base priority to 0x50, blocking the TIM6 system clock interrupt. Even if I set it to 0 after the malloc and before the 500ms delay, it reverts to 0x50 again by the time the delay finishes.

Something is triggering pvPortMalloc without me knowing and something is increasing the base priority. I don't know what at the moment.